First pass at removing NaN. I don't love it. Lots of changes, and hard to test. Shelfing it for now to work on lower hanging fruit.

+++ src/main/java/org/openpnp/machine/reference/ReferenceNozzleTip.java
@@ -124,7 +124,7 @@ public class ReferenceNozzleTip extends AbstractNozzleTip {
                  * everything.
                  */
                 if (changerMidLocation2 == null) {
-                    changerMidLocation2 = changerMidLocation.derive(null, null, null, null);
+                    changerMidLocation2 = changerMidLocation.derive((Double) null, null, null, null);
                 }
                 /*
                  * Backwards compatibility for speed settings.
